# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "description": "错误响应的统一形状——所有域错误都长这样（`{\"detail\": \"一句话\"}`）。\n\n各端点在 `responses=` 里引用它，把错误路径写进 OpenAPI 契约：\n前端生成类型时才能覆盖错误体，而不是只看到成功路径。",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromptCreated": {
        "description": "POST /api/prompts 的响应体（服务端分配 ID）。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "description": "提示词 ID（内部稳定身份，不随改名变化）",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "显示名（可改、允许重名）",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "PromptCreated",
        "type": "object"
      },
      "PromptFull": {
        "description": "提示词全文。",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "description": "提示词 ID（内部稳定身份，不随改名变化）",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "显示名（可改、允许重名）",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "body"
        ],
        "title": "PromptFull",
        "type": "object"
      },
      "PromptInfo": {
        "description": "提示词条目（列表项）。",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "description": "提示词 ID（内部稳定身份，不随改名变化）",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "description": "显示名（可改、允许重名）",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "PromptInfo",
        "type": "object"
      },
      "PromptRenameRequest": {
        "description": "POST /api/prompts/{id}/rename 的请求体。",
        "properties": {
          "new_name": {
            "description": "新显示名（可改、允许重名）",
            "title": "New Name",
            "type": "string"
          }
        },
        "required": [
          "new_name"
        ],
        "title": "PromptRenameRequest",
        "type": "object"
      },
      "PromptSaveRequest": {
        "description": "POST /api/prompts 与 PUT /api/prompts/{id} 的请求体。",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "name": {
            "default": "",
            "description": "显示名；缺省 = 沿用现有显示名",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "PromptSaveRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "AI 打标工具（发图 + 指令产出 caption，支持迭代改写）",
    "title": "Dataset Factory",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/prompts": {
      "get": {
        "description": "列出全部提示词。",
        "operationId": "list_all_api_prompts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PromptInfo"
                  },
                  "title": "Response List All Api Prompts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "提示词库"
        ]
      },
      "post": {
        "description": "新建一条提示词（服务端分配 ID）。",
        "operationId": "create_api_prompts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "提示词库"
        ]
      }
    },
    "/api/prompts/{pid}": {
      "delete": {
        "description": "删除提示词。",
        "operationId": "remove_api_prompts__pid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "提示词不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove",
        "tags": [
          "提示词库"
        ]
      },
      "get": {
        "description": "读某条提示词全文。",
        "operationId": "get_one_api_prompts__pid__get",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptFull"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "提示词文件损坏（格式非法）"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "提示词不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get One",
        "tags": [
          "提示词库"
        ]
      },
      "put": {
        "description": "覆盖保存提示词全文（旧版进 _history 滚动备份；显示名一并更新）。",
        "operationId": "save_api_prompts__pid__put",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "显示名不合法"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "提示词不存在"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "提示词超出 32 KiB 字节护栏"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save",
        "tags": [
          "提示词库"
        ]
      }
    },
    "/api/prompts/{pid}/rename": {
      "post": {
        "description": "改显示名（只写 frontmatter 的 name 字段；引用存 ID、不受影响）。",
        "operationId": "rename_api_prompts__pid__rename_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "新名称不合法"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "要改名的提示词不存在"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename",
        "tags": [
          "提示词库"
        ]
      }
    },
...
  }
}